+++ Windows/WindowsHost.cpp
@@ -87,10 +87,10 @@ WindowsHost::WindowsHost(HWND mainWindow) {
 	//add first XInput device to respond
 	input.push_back(std::shared_ptr<InputDevice>(new XinputDevice()));
 	//find all connected DInput devices of class GamePad
-	int numDInputDevs = DinputDevice::getNumPads();
-	for (int i = 0; i < numDInputDevs; i++)
+	size_t numDInputDevs = DinputDevice::getNumPads();
+	for (size_t i = 0; i < numDInputDevs; i++)
 	{
-		input.push_back(std::shared_ptr<InputDevice>(new DinputDevice(i)));
+		input.push_back(std::shared_ptr<InputDevice>(new DinputDevice(static_cast<int>(i))));
 	}
 	keyboard = std::shared_ptr<KeyboardDevice>(new KeyboardDevice());
 	input.push_back(keyboard);
